fix: Leak in nis-initgroups.c:_nss_nis_initgroups_dyn()
Fix the memory leak in function 'nis-initgroups.c:_nss_nis_initgroups_dyn'. This is done by making the 'tmpbuf'  initialize to NULL and then freeing it in the 'done'. We also need to fix the possible memory leak on 'realloc' when the function 'realloc' fails with the check. This will leak as the previous pointer 'tmpbuf' was called 'realloc' on, and the pointer with the memory was lost (not freed), and the pointer is set to NULL. We introduce a new variable that we check before the assignment.